# docs/plans/artifacts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SignalConstructionBacktestRequest": {
        "description": "Run signal-construction backtest.",
        "properties": {
          "buy_threshold": {
            "default": 0.2,
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Buy Threshold",
            "type": "number"
          },
          "indicators": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Indicators",
            "type": "object"
          },
          "interval": {
            "default": "1d",
            "title": "Interval",
            "type": "string"
          },
          "max_symbols": {
            "default": 100,
            "maximum": 1000.0,
            "minimum": 5.0,
            "title": "Max Symbols",
            "type": "integer"
          },
          "period": {
            "default": "6mo",
            "title": "Period",
            "type": "string"
          },
          "sell_threshold": {
            "default": -0.2,
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Sell Threshold",
            "type": "number"
          },
          "top_n": {
            "default": 20,
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Top N",
            "type": "integer"
          },
          "universe": {
            "default": "XU100",
            "title": "Universe",
            "type": "string"
          }
        },
        "title": "SignalConstructionBacktestRequest",
        "type": "object"
      },
      "SignalConstructionFiveFactorRequest": {
        "description": "Run a five-factor rotation backtest via core service.",
        "properties": {
          "end_date": {
            "default": "2025-12-31",
            "title": "End Date",
            "type": "string"
          },
          "factor_name": {
            "default": "five_factor_rotation",
            "title": "Factor Name",
            "type": "string"
          },
          "max_position_weight": {
            "default": 0.25,
            "maximum": 1.0,
            "minimum": 0.05,
            "title": "Max Position Weight",
            "type": "number"
          },
          "rebalance_frequency": {
            "default": "monthly",
            "title": "Rebalance Frequency",
            "type": "string"
          },
          "start_date": {
            "default": "2020-01-01",
            "title": "Start Date",
            "type": "string"
          },
          "top_n": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 5.0,
            "title": "Top N",
            "type": "integer"
          }
        },
        "title": "SignalConstructionFiveFactorRequest",
        "type": "object"
      },
      "SignalConstructionOrthogonalizationRequest": {
        "description": "Orthogonalization diagnostics request.",
        "properties": {
          "axes": {
            "items": {
              "type": "string"
            },
            "title": "Axes",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "epsilon": {
            "default": 1e-08,
            "exclusiveMinimum": 0.0,
            "title": "Epsilon",
            "type": "number"
          },
          "min_overlap": {
            "default": 20,
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Min Overlap",
            "type": "integer"
          }
        },
        "title": "SignalConstructionOrthogonalizationRequest",
        "type": "object"
      },
      "SignalConstructionSnapshotRequest": {
        "description": "Run signal-construction snapshot.",
        "properties": {
          "buy_threshold": {
            "default": 0.2,
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Buy Threshold",
            "type": "number"
          },
          "indicators": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Indicators",
            "type": "object"
          },
          "interval": {
            "default": "1d",
            "title": "Interval",
            "type": "string"
          },
          "max_symbols": {
            "default": 100,
            "maximum": 1000.0,
            "minimum": 5.0,
            "title": "Max Symbols",
            "type": "integer"
          },
          "period": {
            "default": "6mo",
            "title": "Period",
            "type": "string"
          },
          "sell_threshold": {
            "default": -0.2,
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Sell Threshold",
            "type": "number"
          },
          "top_n": {
            "default": 20,
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Top N",
            "type": "integer"
          },
          "universe": {
            "default": "XU100",
            "title": "Universe",
            "type": "string"
          }
        },
        "title": "SignalConstructionSnapshotRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BIST Quant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/signal-construction/backtest": {
      "post": {
        "description": "Run multi-indicator signal backtest.",
        "operationId": "signal_construction_backtest_api_signal_construction_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalConstructionBacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Signal Construction Backtest Api Signal Construction Backtest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signal Construction Backtest",
        "tags": [
          "signal-construction"
        ]
      }
    },
    "/api/signal-construction/five-factor": {
      "post": {
        "description": "Run five-factor rotation backtest via core service.",
        "operationId": "signal_construction_five_factor_api_signal_construction_five_factor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalConstructionFiveFactorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Signal Construction Five Factor Api Signal Construction Five Factor Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signal Construction Five Factor",
        "tags": [
          "signal-construction"
        ]
      }
    },
    "/api/signal-construction/orthogonalization": {
      "post": {
        "description": "Return orthogonalization configuration diagnostics for selected axes.",
        "operationId": "signal_construction_orthogonalization_api_signal_construction_orthogonalization_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalConstructionOrthogonalizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Signal Construction Orthogonalization Api Signal Construction Orthogonalization Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signal Construction Orthogonalization",
        "tags": [
          "signal-construction"
        ]
      }
    },
    "/api/signal-construction/snapshot": {
      "post": {
        "description": "Run multi-indicator signal snapshot.",
        "operationId": "signal_construction_snapshot_api_signal_construction_snapshot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalConstructionSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Signal Construction Snapshot Api Signal Construction Snapshot Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signal Construction Snapshot",
        "tags": [
          "signal-construction"
        ]
      }
    }
  }
}